Moved class files to a sepatate folder
